+++ src/pages/Buying/Buying.jsx
@@ -6,101 +6,9 @@ import LikeImg from '../../assets/imgs/Funding/LikeImage.svg';
 import Tomato from '../../assets/imgs/Funding/Tomato.png';
 import Plum from '../../assets/imgs/Funding/Plum.png';
 import Peach from '../../assets/imgs/Funding/Peach.png';
+import Items from './Items';
 export default function Buying() {
-  const [products, setProducts] = useState([]);
-
-  useEffect(() => {
-    // api추가
-
-    const fakeProductData = [
-      {
-        id: 1,
-        number: '선택 1',
-        details: '#멋쟁이 토마토',
-        name: '토마토',
-        weight: '10kg',
-        price: '10,000',
-        quantity: 0,
-        imageUrl: Tomato,
-      },
-      {
-        id: 2,
-        number: '선택 2',
-        details: '#멋쟁이 매실',
-        name: '매실',
-        weight: '10kg',
-        price: '20,000',
-        quantity: 0,
-        imageUrl: Plum,
-      },
-      {
-        id: 3,
-        number: '선택 3',
-        details: '#멋쟁이 자두',
-        name: '자두',
-        weight: '10kg',
-        price: '30,000',
-        quantity: 0,
-        imageUrl: Peach,
-      },
-      {
-        id: 3,
-        number: '선택 3',
-        details: '#멋쟁이 자두',
-        name: '자두',
-        weight: '10kg',
-        price: '30,000',
-        quantity: 0,
-        imageUrl: Peach,
-      },
-      {
-        id: 3,
-        number: '선택 3',
-        details: '#멋쟁이 자두',
-        name: '자두',
-        weight: '10kg',
-        price: '30,000',
-        quantity: 0,
-        imageUrl: Peach,
-      },
-      {
-        id: 3,
-        number: '선택 3',
-        details: '#멋쟁이 자두',
-        name: '자두',
-        weight: '10kg',
-        price: '30,000',
-        quantity: 0,
-        imageUrl: Peach,
-      },
-      // 추가가능
-    ];
-
-    setProducts(fakeProductData);
-  }, []);
-
-  const handleQuantityChange = (product, setProducts) => (change) => {
-    // 수량 변경 이벤트 핸들러
-    const newQuantity = Math.max(0, product.quantity + change); // 최소값은 0으로 설정
-
-    setProducts((prevProducts) => {
-      const updatedProducts = prevProducts.map((p) => {
-        if (p.id === product.id) {
-          return { ...p, quantity: newQuantity };
-        }
-        return p;
-      });
-      return updatedProducts;
-    });
-  };
-
-  const handleSelectClick = (product) => {
-    // 담기 버튼 클릭 이벤트 핸들러
-    // 서버로 수량 정보 전송
-    console.log(`Product ${product.name} 수량: ${product.quantity}`);
-    alert(product.name + ' ' + product.quantity + '개를 담았어요.');
-  };
-
+  
   return (
     <>
       <Wrapper>
@@ -135,52 +43,8 @@ export default function Buying() {
         <HeadLine></HeadLine>
         <ProductTitle>상품 종류</ProductTitle>
         <ProductContainer>
-          {products.length > 0 && (
-            <ProductList>
-              {products.map((product) => (
-                <Product key={product.id}>
-                  <img
-                    src={product.imageUrl}
-                    alt={`상품 이미지 - ${product.name}`}
-                  />
-                  <OptionContainer>
-                    <ProductInfo>
-                      {product.number} <a>{product.details}</a>
-                    </ProductInfo>
-                    <ProductName>
-                      품 명 <b>{product.name}</b>
-                    </ProductName>
-                    <ProductWeight>
-                      무 게 <b>{product.weight}</b>
-                    </ProductWeight>
-                    <ProductPrice>
-                      가 격<b>{product.price}원</b>
-                    </ProductPrice>
-                    <QuantityControl>
-                      <QuantityButton
-                        onClick={() =>
-                          handleQuantityChange(product, setProducts)(-1)
-                        }
-                      >
-                        -
-                      </QuantityButton>
-                      <ProductQuantity>{product.quantity}</ProductQuantity>
-                      <QuantityButton
-                        onClick={() =>
-                          handleQuantityChange(product, setProducts)(1)
-                        }
-                      >
-                        +
-                      </QuantityButton>
-                    </QuantityControl>
-                    <SelectButton onClick={() => handleSelectClick(product)}>
-                      담기
-                    </SelectButton>
-                  </OptionContainer>
-                </Product>
-              ))}
-            </ProductList>
-          )}
+          <Items/> 
+          {/* 상품 컴포넌트 */}
         </ProductContainer>
         <HeadLine></HeadLine>
         <ProductTitle>상품 소개</ProductTitle>
@@ -362,148 +226,6 @@ const ProductContainer = styled.div`
   overflow-x: auto;
   overflow-y: auto; // 세로 스크롤이 생기지 않도록 설정
   white-space: nowrap; // 가로 스크롤을 통해 한 줄에 표시
-`;
-
-const Product = styled.div`
-  display: flex;
-  flex-direction: column;
-  position: relative;
-  margin: 40px 20px;
-  width: 400px; // 상품 박스의 너비 조정
-  height: 530px;
-  background: #fff;
-  border-radius: 10px;
-  filter: drop-shadow(0px 1px 15px rgba(0, 0, 0, 0.2));
-  img {
-    width: 100%;
-    height: 50%;
-    object-fit: cover;
-    border-radius: 10px 10px 0 0;
-  }
-`;
-const ProductList = styled.div`
-  display: flex;
-  flex-wrap: nowrap; // 줄 바꿈 방지
-`;
-
-const OptionContainer = styled.div`
-  display: flex;
-  width: 358px;
-  padding: 32px;
-  flex-direction: column;
-  align-items: flex-start;
-  b {
-    color: #000;
-    font-family: Pretendard;
-    font-size: 20px;
-    font-style: normal;
-    font-weight: 400;
-    line-height: normal;
-    margin-left: 15px;
-  }
-`;
-
-const ProductInfo = styled.div`
-  color: #000;
-  font-family: Pretendard;
-  font-size: 28px;
-  font-style: normal;
-  font-weight: 600;
-  line-height: normal;
-  a {
-    color: var(--Point-Color, #ff9c2f);
-    font-family: Pretendard;
-    font-size: 20px;
-    font-style: normal;
-    font-weight: 600;
-    line-height: normal;
-    margin-left: 14px;
-  }
-`;
-
-const ProductName = styled.div`
-  color: #000;
-  font-family: Pretendard;
-  font-size: 20px;
-  font-style: normal;
-  font-weight: 600;
-  line-height: normal;
-  margin-top: 30px;
-`;
-
-const ProductWeight = styled.div`
-  color: #000;
-  font-family: Pretendard;
-  font-size: 20px;
-  font-style: normal;
-  font-weight: 600;
-  line-height: normal;
-  margin-top: 10px;
-`;
-
-const ProductPrice = styled.div`
-  color: #000;
-  font-family: Pretendard;
-  font-size: 20px;
-  font-style: normal;
-  font-weight: 600;
-  line-height: normal;
-  margin-top: 10px;
-`;
-
-const QuantityControl = styled.div`
-  display: flex;
-  width: 116.215px;
-  height: 46px;
-  flex-shrink: 0;
-  align-items: center;
-  margin-top: 10px;
-  border-radius: 84px;
-  border: 1px solid #969696;
-  background: #fff;
-  justify-content: space-between;
-`;
-
-const QuantityButton = styled.button`
-  width: 40px;
-  height: 30px;
-  border: none;
-  border-radius: 5px;
-  text-align: center;
-  font-family: Pretendard;
-  font-size: 23px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: normal;
-  stroke: var(--Font-Color, #2c2c2c);
-  padding-bottom: 3px;
-`;
-
-const ProductQuantity = styled.div`
-  font-size: 16px;
-  font-weight: bold;
-`;
-const SelectButton = styled.div`
-  width: 143px;
-  display: flex;
-  position: relative;
-  left: 11rem;
-  top: -3rem;
-  height: 46px;
-  flex-shrink: 0;
-  justify-content: center;
-  align-items: center;
-  border-radius: 84px;
-  border: 1px solid var(--Main-Color, #ff4256);
-  background: #ff4257;
-  color: #fff;
-  text-align: center;
-  font-family: Pretendard;
-  font-size: 20px;
-  font-style: normal;
-  font-weight: 600;
-  line-height: normal;
-  cursor: pointer;
 `;
 
 const InfoContainer = styled.div`
